+++ sdk-integration-tests/src/main/java/com/kount/ris/util/Utilities.java
@@ -17,7 +17,7 @@ public final class Utilities {
 	private static final Address SHIPPING_ADDRESS = 
 			new Address("567 West S2A1 Court North", null, "Gnome", "AK", "99762", "US");
 	
-	public static Inquiry defaultInquiry(String sessionId, int merchantId) {
+	public static Inquiry defaultInquiry(String sessionId, long merchantId) {
 		Inquiry result = new Inquiry();
 		// the value here gets khashed automatically by the setter method request.setPayment(p)
 		Payment payment = new CardPayment("0007380568572514");
